doc: update README and manpages
* fix: pen sprite naming

+++ include/image_loader/bongocat/load_images_bongocat.h
@@ -18,7 +18,7 @@ load_bongocat_anim(int anim_index, get_sprite_callback_t get_sprite, size_t embe
                    load_bongocat_anim_type_t type, anim_sprite_sheet_from_embedded_svgs_t svg_params,
                    anim_sprite_sheet_from_embedded_svgs_cropping_t cropping);
 
-bongocat_error_t init_bongocat_anim(animation_thread_context_t& ctx, int anim_index, get_sprite_callback_t get_sprite,
+bongocat_error_t init_bongocat_anim(animation_thread_context_t& ctx, size_t anim_index, get_sprite_callback_t get_sprite,
                                     size_t embedded_images_count, load_bongocat_anim_type_t type, anim_sprite_sheet_from_embedded_svgs_t svg_params,
                                     anim_sprite_sheet_from_embedded_svgs_cropping_t cropping);
